Restructure journey command.
